Server/Packets: Fix Arithmetic issue in ReadString.
Server/Packets: Fix Arithmetic issue in ReadString.

+++ src/server/shared/Packets/ByteBuffer.h
@@ -622,7 +622,7 @@ public:
         }
     }
 
-    std::string ReadString(uint32 length)
+    std::string ReadString(size_t length)
     {
         if (!length)
             return std::string();
